Fix issue with Firebase passing an int instead of double for whole numbers.

+++ lib/providers/orders.dart
@@ -48,7 +48,7 @@ class Orders with ChangeNotifier {
               .map(
                 (item) => CartItem(
                   id: item['id'],
-                  price: item['price'],
+                  price: item['price'].toDouble(),
                   quantity: item['quantity'],
                   title: item['title'],
                 ),
